requirements.txt
numpy==1.19.5
matplotlib==3.3.3
attrs==20.3.0
pandas==1.2.0
python-dotenv==0.17.0
seaborn==0.11.1
scipy==1.6.2
tweepy==3.10.0
